requirements.txt
absl-py==0.6.1
APScheduler==3.6.1
asn1crypto==0.24.0
astor==0.7.1
certifi==2018.10.15
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
dlib==19.18.0
Flask==1.1.1
Flask-APScheduler==1.11.0
flask-core==2.9.0
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gast==0.2.0
gevent==1.4.0
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.16.0
h5py==2.8.0
idna==2.6
imageio==2.6.1
imutils==0.4.6
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.14.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.1
mxnet==1.3.0
networkx==2.3
numpy==1.14.6
opencv-contrib-python==3.4.4.19
passlib==1.7.1
Pillow==6.2.0
protobuf==3.6.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
redis==3.3.8
requests==2.18.4
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.3.8
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
torch==1.3.0+cpu
torchvision==0.4.1+cpu
tornado==6.0.3
tzlocal==2.0.0
urllib3==1.22
Werkzeug==0.16.0
WTForms==2.2.1
xlrd==1.2.0
